- externalized data into separate file - Customized look and feel
