+++ src/index.css
@@ -5,8 +5,12 @@ html {
 body {
     margin: 0;
     padding: 0;
+    background-color: #2c3e50;
 }
 
+#content-root {
+    background-color: white;
+}
 /*
  * Author: http://stuffandnonsense.co.uk/blog/about/hardboiled_css3_media_queries/
  */
